loops videos and removes audio-video sync
paula thought that there were some stanzas where the discrepancy between audio and video duration made it too obvious which audio was correct... this made the exercise less about typography and more about finding the audio that syncs... so: video loops and audio doesn't sync automatically. the idea here is that participants may still sync video and audio if they want, but they must do so by pressing play on the audio at the right time. since this is a somewhat imprecise method, they may find that small discrepancies are due to them and not the audio/video length. (we hope.)